Fix GraphQL schema issues and improve error handling for production stability
- Fix scan_duration GraphQL field from Int to Float to handle decimal values
- Improve GPT response parsing with better error handling for undefined/null responses
- Add detailed logging for GPT chunk processing failures
- Add generic HTML/CSS/JavaScript fallback for tech stack API failures
- Remove inappropriate fallbacks for accessibility issues (only tech stack gets fallback)
- Ensure production stability with graceful error handling

+++ api/graphql/schemas/accessibilityReport.schema.ts
@@ -213,7 +213,7 @@ export const AccessibilitySchema = gql`
 
   type WebAbilityMetadata {
     job_id: String!
-    scan_duration: Int!
+    scan_duration: Float!
     scan_timestamp: String!
     browser_info: WebAbilityBrowserInfo!
     accessibility_standards: WebAbilityAccessibilityStandards!
